Zmodyfikowano kategorie, ćwiczenia i dodano nazwe do składnika

+++ WebApplication/Controllers/CwiczenieController.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +12,7 @@ using WebApplication.Models;
 
 namespace WebApplication.Controllers
 {
+    [Authorize]
     public class CwiczenieController : Controller
     {
         private readonly MyContext _context;
@@ -48,6 +51,9 @@ namespace WebApplication.Controllers
         // GET: Cwiczenie/Create
         public IActionResult Create()
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             ViewData["id_kategorii"] = new SelectList(_context.kategoriaCwiczenia, "id_kategorii", "nazwa");
             return View();
         }
@@ -59,6 +65,9 @@ namespace WebApplication.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("id_cwiczenia,nazwa,opis,spalone_kalorie,id_kategorii")] Cwiczenie cwiczenie)
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             if (ModelState.IsValid)
             {
                 _context.Add(cwiczenie);
@@ -72,6 +81,9 @@ namespace WebApplication.Controllers
         // GET: Cwiczenie/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             if (id == null)
             {
                 return NotFound();
@@ -93,6 +105,9 @@ namespace WebApplication.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("id_cwiczenia,nazwa,opis,spalone_kalorie,id_kategorii")] Cwiczenie cwiczenie)
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             if (id != cwiczenie.id_cwiczenia)
             {
                 return NotFound();
@@ -123,8 +138,12 @@ namespace WebApplication.Controllers
         }
 
         // GET: Cwiczenie/Delete/5
+
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             if (id == null)
             {
                 return NotFound();
@@ -146,6 +165,9 @@ namespace WebApplication.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!this.isTrainer())
+                return RedirectToAction("Index");
+
             var cwiczenie = await _context.cwiczenia.FindAsync(id);
             _context.cwiczenia.Remove(cwiczenie);
             await _context.SaveChangesAsync();
@@ -155,6 +177,17 @@ namespace WebApplication.Controllers
         private bool CwiczenieExists(int id)
         {
             return _context.cwiczenia.Any(e => e.id_cwiczenia == id);
+        }
+
+        private bool isTrainer()
+        {
+            int userId = int.Parse(User.Identity.GetUserId());
+            List<RolaUzytkownika> usersRoles = _context.RolaUzytkownika.Where(k => k.id_uzytkownika == userId).Include(c => c.rola).ToList();
+
+            foreach (var usersRole in usersRoles)
+                if (usersRole.rola.nazwa == "trener" || usersRole.rola.nazwa == "admin")
+                    return true;
+            return false;
         }
     }
 }
